requirements.txt.txt
Flask
numpy
tensorflow
pandas
jieba
google-genai
requests # (雖然是測試用，但最好包含)
gunicorn # (這是生產環境中替代 Flask 內建伺服器的工具)